revert config to only contains ExchangeConfig
