+++ go.mod
@@ -7,7 +7,8 @@ require (
 	github.com/google/go-querystring v1.1.0
 	github.com/gorilla/websocket v1.4.2
 	github.com/openware/pkg v0.0.0-20210528154413-404a657867b9
+	github.com/stretchr/testify v1.7.0
 	github.com/toorop/go-pusher v0.0.0-20180521062818-4521e2eb39fb
-	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9
+	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9 // indirect
 	golang.org/x/time v0.0.0-20210220033141-f8bda1e9f3ba
 )
